pass quality to streamlink (#25)
if 720p60 is chosen, we use bitrate as the key and choose the best
stream, as it'll have the highest bitrate. If 720p60 isn't available,
and since best is used, it'll default to the next highest bitrate. All
other quality settings, if chosen, match exactly. So if it isn't found,
it'll fail. If no quality is chosen, best is always used.