Sort games for consistent output
The API changes how games are sorted in the response based on the game
status (In Progress, Scheduled, Final). Sort the games by their datetime
then away team name for consistent output in the Select and Generate
outputs. Also use BTreeMap instead of Hashmap for the games streams so
it retains its original order.